template type for array size size_t rather than int

+++ src/inc/global.hpp
@@ -10,7 +10,7 @@ using complex = std::complex<double>;
 using view_1d = boost::multi_array<complex, 2>::array_view<1>::type;
 using view_2d = boost::multi_array<complex, 3>::array_view<2>::type;
 using range = boost::multi_array_types::index_range;
-template <typename T, int N> struct m_array {
+template <typename T, size_t N> struct m_array {
 	boost::multi_array<T, N> data;
 	std::string name;
 	bool calculated;
@@ -33,7 +33,7 @@ namespace fft {
 }
 
 namespace tools {
-	template <typename T, int N> void multiply_array(m_array<T, N>& field, const double scalar);
+	template <typename T, size_t N> void multiply_array(m_array<T, N>& field, const double scalar);
 	std::vector<complex> array_to_vec(view_1d view);
 	std::vector<complex> array_to_vec(view_2d view);
 	void vec_to_array(view_1d& view, std::vector<complex> vec);
@@ -42,7 +42,7 @@ namespace tools {
 	std::vector<double> get_real(m_array<complex, 3> field, std::array<int, 3> size);
 }
 
-template <typename T, int N>
+template <typename T, size_t N>
 void tools::multiply_array(m_array<T, N>& field, const double scalar) {
 	for (size_t i = 0; i < field.data.num_elements(); i++) {
 		field.data.data()[i] *= scalar;
